add more warnings, correct code according them

+++ src/memory.h
@@ -62,10 +62,10 @@ IS_EXTERN uint32_t reg_gp[REG_GP_COUNT];
 IS_EXTERN uint32_t reg_pc[REG_PC_COUNT];
 #undef IS_EXTERN
 
-int mem_init(char const *mem_path);
-int reg_init(char const *reg_path);
-int mem_save(char const *mem_path);
-int reg_save(char const *reg_path);
+ssize_t mem_init(char const *mem_path);
+ssize_t reg_init(char const *reg_path);
+ssize_t mem_save(char const *mem_path);
+ssize_t reg_save(char const *reg_path);
 
 int mem_print(size_t size);
 int reg_print();
